vision-service/requirements.txt
# vision-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
torch==2.2.0
torchvision==0.17.0
python-chess==1.999
werkzeug==3.0.1
